requirements.txt
# dev dependencies
mock                             # mock objects for unit tests
sphinx                           # doc generator

# kubernetes deployment dependencies
Jinja2==2.9.6                    # jinja templates are used in the kubernetes deployment scripts
PyYAML==3.12                     # yaml parsing

# runtime dependencies
beautifulsoup4==4.5.3            # used for html parsing utilites
Django==1.11                     # core server-side framework
django-guardian==1.4.6           # object-level permissions for database records
django-activity-stream==0.6.3    # used to log various user and application actions
django-hijack==2.1.3             # allows admins to login as other user
ipython                          # used by shell.py
numpy==1.12.1
openpyxl==2.4.8                  # library for reading/writing Excel files
xlrd==1.0.0                      # faster library for parsing Excel
pandas==0.19.2
psycopg2==2.7.1                  # postgres database access
PyYAML==3.12                     # yaml parsing
requests==2.13.0                 # simpler way to make http requests
tqdm                             # convenient way to create progress bar for long-running command-line operations
whitenoise==3.3.0                # simplified static file handling

# legacy dependencies we're moving away from
Pillow==4.0.0                    # required dependency of Djagno ImageField-type database records
pysam==0.10.0                    # bam and tabix library
django-pretty-times==0.2.0       # date rendering should happen in javascript
django-compressor==2.1.1
pymongo==3.4.0                   # switching to other database
PyVCF==0.6.8
Markdown==2.6.8
slacker==0.9.42                  # intend to move this to separate external modules
django-crispy-forms==1.6.1       # switching to client-side form creation/rendering
progressbar==2.3                 # using tqdm instead

#django-appconf

#selenium
#sh
#simplejson
#reportlab
